Check for missing download_url in index results.
Not having this check was harmless, but led to surprising behavior.

If the response happened not to contain a "download_url" field, then the
ooniResult struct would store "" in the DownloadURL field. url.Parse("")
doesn't return an error, but just an empty string in every field,
including Path. Path is what the program uses when testing to see
whether a report was already downloaded. Because
filepath.join(outputDirectory, "") is just outputDirectory, this would
cause messages like
	 1/11 exists: .
	 2/11 exists: .
	 3/11 exists: .
	 4/11 exists: .
	...

Indexes never lack the "download_url" in practice. I only found it by
accident when experimenting with the /api/v1/measurements endpoint,
which calls the corresponding field "measurement_url" instead of
"download_url".